+++ include/anre/RelationshipEngine.hpp
@@ -9,9 +9,6 @@ namespace anre {
 class RelationshipEngine {
 public:
     std::vector<EventNode> buildGraph(const std::vector<SecurityEvent>& events);
-
-private:
-    static std::int64_t findNodeByPid(const std::vector<EventNode>& nodes, std::int32_t pid);
 };
 
 } // namespace anre